Revert "applied seperation of concerns"
