--- CodeSmellsRefactorExercise/Interface/ICalculator.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace CodeSmellsRefactorExercise.Interface;
-
-public interface ICalculator
-{
-    public void Run();
-}
