--- CodeSmellsRefactorExercise/Interface/IParse.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace CodeSmellsRefactorExercise.Interface;
-
-public interface IParse
-{
-    public List<int> ParseToList(string input, List<int> numbers);
-}
