+++ CodeSmellsRefactorExercise/Program.cs
@@ -1,14 +1,50 @@
 ﻿namespace CodeSmellsRefactorExercise
 {
-	public class Program
+	internal class Program
 	{
 		static void Main(string[] args)
 		{
-			var parse = new Parse();
-			var operation = new Operation();
-			var calculator = new Calculator(operation, parse);
+			var calculator = new Calculator();
 			calculator.Run();
 		}
 	}
 
+	public class Calculator
+	{
+		private List<int> numbers = new List<int>();
+
+		public void Run()
+		{
+			Console.WriteLine("Enter numbers separated by spaces:");
+			var input = Console.ReadLine();
+
+			if (string.IsNullOrEmpty(input))
+			{
+				Console.WriteLine("No input provided.");
+				return;
+			}
+
+			var numList = input.Split(' ');
+			foreach (var n in numList)
+			{
+				numbers.Add(int.Parse(n));
+			}
+
+			var sum = 0;
+			foreach (var number in numbers)
+			{
+				sum += number;
+			}
+
+			var average = sum / numbers.Count;
+
+			if (numbers.Count > 0 && sum > 0 && average > 0)
+			{
+				Console.WriteLine($"Sum: {sum}");
+				Console.WriteLine($"Average: {average}");
+			}
+
+			Console.WriteLine("Program finished.");
+		}
+	}
 }
